requirements.txt
flask-cors==3.0.10
appdirs==1.4.4
black==20.8b1
boto3==1.17.12
botocore>=1.20.12
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
Flask==1.0.2
gunicorn==20.0.4
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
jmespath==0.10.0
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.0
python-dateutil==2.8.1
regex==2020.11.13
requests>=2.21.0
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3>=1.21.1
Werkzeug==0.14.1
flask-migrate==2.6.0
flask-script==2.0.6
direnv==2020.12.3
python-decouple==3.4
psycopg2-binary==2.8.6
Flask==1.0.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Script==2.0.6
